+++ changelog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Updated documentation
 - Added use of retrofit for api calls 
 - Added POJOs
+- Updating IDApi to use SmileIdentityService
 
 ## [0.0.2] - 2019-08-12
 ### Added